Strengthen multiset disj union lemma
This adds a multiset-equivalent to the set API's
`set_fold_disj_union_strong`, which strengthens the fold function
condition to only need to apply to (multi)set elements - note that in a
multiset, elements can be equal, so there is no inequality in the fold
condition. It also now applies to any preorder which respects the
fold function.