Strengthen all fold commutativity lemmas
All these lemmas admit stronger versions which only care about the
commutativity of the fold and accumulator functions for the data
structure elements, rather than for the whole type.

Each lemma has been strengthened in line with the rest of its file,
which is why there are some discrepancies across the definitions.

This was inspired by the fact that the original version of the lemma for
maps was in fact slightly stronger in its commutativity conditions.

Only the map version permits a preorder instead of an equivalence -
though in principle, the commutativity conditions should allow all
versions to only require a preorder. The reason for this is that the map
induction principle used is much stronger, and already handles insertion
commutativity "internally", which allows for the right kind of
rewriting. The other proofs get stuck trying to show that the preorder
should strengthen to an equivalence on a certain class of elements in
the fold. To save my sanity, I've just required an equivalence
everywhere for these instead.